requirements.txt
aiohttp==3.8.0
aiohttp-retry==2.4.6
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.0
asynctest==0.13.0
atpublic==2.3
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cycler==0.11.0
dictdiffer==0.9.0
diskcache==5.2.1
distro==1.6.0
dpath==2.0.5
dulwich==0.20.26
dvc==2.8.2
Flask==2.0.2
Flask-Cors==3.0.10
flatten-dict==0.4.2
flufl.lock==6.0
fonttools==4.28.2
frozenlist==1.2.0
fsspec==2021.10.1
ftfy==6.0.3
funcy==1.16
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
grandalf==0.6
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
kiwisolver==1.3.2
mailchecker==4.0.15
MarkupSafe==2.0.1
matplotlib==3.5.0
multidict==5.2.0
nanotime==0.5.2
networkx==2.6.3
numpy==1.21.4
packaging==21.2
pandas==1.3.4
pathspec==0.9.0
phonenumbers==8.12.36
Pillow==8.4.0
ply==3.11
psutil==5.8.0
pyasn1==0.4.8
pycparser==2.20
pydot==1.4.2
pygit2==1.7.0
Pygments==2.10.0
pygtrie==2.4.2
pyparsing==2.4.7
python-benedict==0.24.3
python-dateutil==2.8.2
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rich==10.12.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
setuptools-scm==6.3.2
shortuuid==1.0.1
shtab==1.4.2
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
-e git+https://github.com/SiddharthaShandilya/air_quality_index_prediction.git@41664bf4e84f4d5482f44a0c926c5000f39059f1#egg=src
tabulate==0.8.9
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
voluptuous==0.12.2
wcwidth==0.2.5
Werkzeug==2.0.2
wincertstore==0.2
xgboost==1.5.1
xmltodict==0.12.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.6.0
